fix(internal): fix version string mangling logic (#16)
Before this assumed that all tags started with `v` and cut off the v:

```go
// current version: v1.2.3
fmt.Println(internal.GitVersion()) // 1.2.3
```

This fails horribly when versions do not start with `v`:

```go
// current version: 1.2.3
fmt.Println(internal.GitVersion()) // .2.3
```

This breaks Debian's package manager[0] horribly because `.2.3` is not a
valid version number in Debian land.

There needs to be additional validation for version numbers pre-build
time, but for now it's easy enough to make sure that this exact failure
case doesn't happen again.

[0] https://github.com/TecharoHQ/anubis/pull/373#issuecomment-2832471884

Signed-off-by: Xe Iaso <me@xeiaso.net>